refactor(goja-repl): split root.go into per-command files (CQ-8)
Split the 568-line root.go into a 205-line shared core (rootOptions,
commandSupport, appSupportOptions, helpers, settings types) plus one
file per CLI command:

  cmd_sessions.go  - sessions list
  cmd_create.go    - create session
  cmd_eval.go      - evaluate source
  cmd_snapshot.go  - session snapshot
  cmd_history.go   - evaluation history
  cmd_bindings.go  - binding inspection
  cmd_docs.go      - REPL-authored docs
  cmd_export.go    - session export
  cmd_restore.go   - replay-restore
  cmd_serve.go     - HTTP server

Also extracted runWithApp() helper to eliminate the repeated
newApp/defer-Close/call-method pattern across commands.

GOJA-044 CQ-8
