+++ cmd/goja-repl/cmd_export.go
@@ -1,0 +1,43 @@
+package main
+
+import (
+	"io"
+	"context"
+
+	"github.com/go-go-golems/glazed/pkg/cmds"
+	"github.com/go-go-golems/glazed/pkg/cmds/fields"
+	"github.com/go-go-golems/glazed/pkg/cmds/schema"
+	"github.com/go-go-golems/go-go-goja/pkg/replapi"
+	"github.com/go-go-golems/glazed/pkg/cmds/values"
+)
+
+type exportCommand struct {
+	*cmds.CommandDescription
+	commandSupport
+}
+
+var _ cmds.BareCommand = (*exportCommand)(nil)
+
+func newExportCommand(out io.Writer, opts *rootOptions) *exportCommand {
+	return &exportCommand{
+		CommandDescription: cmds.NewCommandDescription("export",
+			cmds.WithShort("Export a session from SQLite"),
+			cmds.WithFlags(fields.New("session-id", fields.TypeString, fields.WithRequired(true), fields.WithHelp("Persistent session id"))),
+		),
+		commandSupport: commandSupport{out: out, opts: opts},
+	}
+}
+
+func (c *exportCommand) Run(ctx context.Context, vals *values.Values) error {
+	settings := sessionSettings{}
+	if err := vals.DecodeSectionInto(schema.DefaultSlug, &settings); err != nil {
+		return err
+	}
+	return c.runWithApp(func(ctx context.Context, app *replapi.App) error {
+		exported, err := app.Export(ctx, settings.SessionID)
+		if err != nil {
+			return err
+		}
+		return writeJSON(c.out, exported)
+	})(ctx, vals)
+}
